+++ src/development/vortex/development/core/engine/trainer/lr_scheduler/cosine_lr_with_warmup.py
@@ -9,9 +9,9 @@ from .scheduler import Scheduler
 logger = logging.getLogger(__name__)
 
 
-class CosineLRScheduler(Scheduler):
+class CosineLRWithWarmUp(Scheduler):
     """
-    Cosine decay with restarts.
+    Cosine decay with warm up and restarts.
     This is described in the paper https://arxiv.org/abs/1608.03983.
     Inspiration from
     https://github.com/allenai/allennlp/blob/master/allennlp/training/learning_rate_schedulers/cosine.py